+++ nengo/neurons.py
@@ -471,6 +471,57 @@ class Sigmoid(NeuronType):
         rates[...] = (1.0 / self.tau_ref) / (1 + np.exp(-J))
 
 
+class Tanh(NeuronType):
+    """A non-spiking neuron model whose response curve is a hyperbolic tangent.
+
+    Parameters
+    ----------
+    tau_ref : float
+        The neuron refractory period, in seconds. The maximum firing rate of the
+        neurons is ``1 / tau_ref``. Must be positive (i.e. ``tau_ref > 0``).
+    initial_state : {str: Distribution or array_like}
+        Mapping from state variables names to their desired initial value.
+        These values will override the defaults set in the class's state attribute.
+    """
+
+    state = {"rates": Choice([0])}
+
+    tau_ref = NumberParam("tau_ref", low=0, low_open=True)
+
+    def __init__(self, tau_ref=0.0025, initial_state=None):
+        super().__init__(initial_state)
+        self.tau_ref = tau_ref
+
+    def gain_bias(self, max_rates, intercepts):
+        """Analytically determine gain, bias."""
+        max_rates = np.array(max_rates, dtype=float, copy=False, ndmin=1)
+        intercepts = np.array(intercepts, dtype=float, copy=False, ndmin=1)
+
+        inv_tau_ref = 1.0 / self.tau_ref
+        if not np.all(max_rates < inv_tau_ref):
+            raise ValidationError(
+                "Max rates must be below the inverse refractory period (%0.3f)"
+                % inv_tau_ref,
+                attr="max_rates",
+                obj=self,
+            )
+
+        inverse = np.arctanh(max_rates * self.tau_ref)
+        gain = inverse / (1.0 - intercepts)
+        bias = -gain * intercepts
+        return gain, bias
+
+    def max_rates_intercepts(self, gain, bias):
+        """Compute the inverse of gain_bias."""
+        intercepts = -bias / gain
+        max_rates = (1.0 / self.tau_ref) * np.tanh(gain + bias)
+        return max_rates, intercepts
+
+    def step(self, dt, J, rates):
+        """Implement the tanh nonlinearity."""
+        rates[...] = (1.0 / self.tau_ref) * np.tanh(J)
+
+
 class LIFRate(NeuronType):
     """Non-spiking version of the leaky integrate-and-fire (LIF) neuron model.