+++ nengo/solvers.py
@@ -170,7 +170,7 @@ class LstsqNoise(Solver):
 
     def __call__(self, A, Y, rng=np.random):
         tstart = time.time()
-        sigma = self.noise * A.max()
+        sigma = self.noise * np.amax(np.abs(A))
         A = A + rng.normal(scale=sigma, size=A.shape)
         X, info = self.solver(A, Y, 0, rng=rng)
         info["time"] = time.time() - tstart
@@ -242,7 +242,7 @@ class LstsqL2nz(LstsqL2):
         # Compute the equivalent noise standard deviation. This equals the
         # base amplitude (noise_amp times the overall max activation) times
         # the square-root of the fraction of non-zero components.
-        sigma = (self.reg * A.max()) * np.sqrt((A > 0).mean(axis=0))
+        sigma = (self.reg * np.amax(np.abs(A))) * np.sqrt((np.abs(A) > 0).mean(axis=0))
 
         # sigma == 0 means the neuron is never active, so won't be used, but
         # we have to make sigma != 0 for numeric reasons.